fix err

+++ pkg/common/hashmap/joinmap.go
@@ -21,7 +21,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/pb/plan"
 )
 
-func NewJoinMap(sels [][]int32, nullSels []int64, expr *plan.Expr, mp *StrHashMap, hasNull bool, idx *index.LowCardinalityIndex) *JoinMap {
+func NewJoinMap(sels [][]int32, nullSels []int32, expr *plan.Expr, mp *StrHashMap, hasNull bool, idx *index.LowCardinalityIndex) *JoinMap {
 	cnt := int64(1)
 	return &JoinMap{
 		cnt:      &cnt,
@@ -54,7 +54,7 @@ func (jm *JoinMap) Index() *index.LowCardinalityIndex {
 	return jm.idx
 }
 
-func (jm *JoinMap) Nullsels() []int64 {
+func (jm *JoinMap) Nullsels() []int32 {
 	return jm.nullSels
 }
